+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/JibBuildRunner.java
@@ -192,6 +192,7 @@ public class JibBuildRunner {
   private final HelpfulSuggestions helpfulSuggestions;
   @Nullable private Path imageDigestOutputPath;
   @Nullable private Path imageIdOutputPath;
+  @Nullable private Path imageJsonOutputPath;
 
   @VisibleForTesting
   JibBuildRunner(
@@ -237,6 +238,11 @@ public class JibBuildRunner {
         String imageId = jibContainer.getImageId().toString();
         Files.write(imageIdOutputPath, imageId.getBytes(StandardCharsets.UTF_8));
       }
+      if (imageJsonOutputPath != null) {
+        ImageMetadataOutput metadataOutput = ImageMetadataOutput.fromJibContainer(jibContainer);
+        String imageJson = metadataOutput.toJson();
+        Files.write(imageJsonOutputPath, imageJson.getBytes(StandardCharsets.UTF_8));
+      }
 
       return jibContainer;
 
@@ -272,8 +278,9 @@ public class JibBuildRunner {
       throw new BuildStepsExecutionException(message, ex);
 
     } catch (ExecutionException ex) {
-      String message = Verify.verifyNotNull(ex.getCause().getMessage()); // keep null-away happy
-      throw new BuildStepsExecutionException(message, ex.getCause());
+      String message = ex.getCause().getMessage();
+      throw new BuildStepsExecutionException(
+          message == null ? "(null exception message)" : message, ex.getCause());
 
     } catch (InterruptedException ex) {
       throw new BuildStepsExecutionException(helpfulSuggestions.none(), ex);
@@ -302,6 +309,18 @@ public class JibBuildRunner {
    */
   public JibBuildRunner writeImageId(@Nullable Path imageIdOutputPath) {
     this.imageIdOutputPath = imageIdOutputPath;
+    return this;
+  }
+
+  /**
+   * Set the location where the image metadata json will be saved. If {@code null} then the metadata
+   * is not saved.
+   *
+   * @param imageJsonOutputPath the location to write the image metadata, or {@code null} to skip
+   * @return this
+   */
+  public JibBuildRunner writeImageJson(@Nullable Path imageJsonOutputPath) {
+    this.imageJsonOutputPath = imageJsonOutputPath;
     return this;
   }
 }